requirements_streamlit_final.txt
streamlit==1.37.0
tensorflow==2.15.0
ultralytics==8.2.50
pillow
opencv-python-headless
torch==2.3.1
torchvision==0.18.1
numpy
